[VirtualInput] Removing old VirtualInput
