+++ src/main/java/com/minecrafttas/tasmod/mixin/playbackhooks/MixinGuiChat.java
@@ -1,5 +1,6 @@
 package com.minecrafttas.tasmod.mixin.playbackhooks;
 
+import com.minecrafttas.tasmod.virtual.VirtualInput;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
@@ -11,7 +12,7 @@ import net.minecraft.client.gui.GuiChat;
 @Mixin(GuiChat.class)
 public class MixinGuiChat {
 	/**
-	 * @return {@link com.minecrafttas.tasmod.virtual.VirtualInput2.VirtualMouseInput#getEventMouseScrollWheel()}
+	 * @return {@link VirtualInput.VirtualMouseInput#getEventMouseScrollWheel()}
 	 */
 	@Redirect(method = "handleMouseInput", at = @At(value = "INVOKE", target = "Lorg/lwjgl/input/Mouse;getEventDWheel()I", remap = false))
 	public int redirectHandleMouseInput4() {
