+++ src/main/java/com/minecrafttas/tasmod/virtual/VirtualCameraAngle.java
@@ -2,27 +2,30 @@ package com.minecrafttas.tasmod.virtual;
 
 import java.io.Serializable;
 
-import com.minecrafttas.tasmod.playback.PlaybackSerialiser;
+public class VirtualCameraAngle implements Serializable {
+	private Float pitch;
+	private Float yaw;
 
-public class VirtualCameraAngle implements Serializable{
-	private float pitch;
-	private float yaw;
 
 	public VirtualCameraAngle() {
-		pitch = 0;
-		yaw = 0;
+		this(null, null);
 	}
 
-	public VirtualCameraAngle(float pitch, float yaw) {
+	public VirtualCameraAngle(Float pitch, Float yaw) {
 		this.pitch = pitch;
 		this.yaw = yaw;
 	}
 
-	public float getPitch() {
+	public void update(Float pitch, Float yaw) {
+		this.pitch = pitch;
+		this.yaw = yaw;
+	}
+	
+	public Float getPitch() {
 		return pitch;
 	}
 
-	public float getYaw() {
+	public Float getYaw() {
 		return yaw;
 	}
 
@@ -30,9 +33,18 @@ public class VirtualCameraAngle implements Serializable{
 	public VirtualCameraAngle clone() {
 		return new VirtualCameraAngle(pitch, yaw);
 	}
-	
+
+	@Override
+	public boolean equals(Object obj) {
+		if (obj instanceof VirtualCameraAngle) {
+			VirtualCameraAngle angle = (VirtualCameraAngle) obj;
+			return pitch == angle.pitch && yaw == angle.yaw;
+		}
+		return super.equals(obj);
+	}
+
 	@Override
 	public String toString() {
-		return PlaybackSerialiser.SectionsV1.CAMERA.getName()+":"+pitch+";"+yaw;
+		return String.format("%s;%s", pitch, yaw);
 	}
 }
