+++ src/test/java/tasmod/virtual/VirtualKeyboardTest.java
@@ -7,9 +7,9 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 
 import org.junit.jupiter.api.Test;
 
-import com.minecrafttas.tasmod.virtual.VirtualKey2;
-import com.minecrafttas.tasmod.virtual.VirtualKeyboard2;
-import com.minecrafttas.tasmod.virtual.VirtualKeyboardEvent;
+import com.minecrafttas.tasmod.virtual.VirtualKey;
+import com.minecrafttas.tasmod.virtual.VirtualKeyboard;
+import com.minecrafttas.tasmod.virtual.event.VirtualKeyboardEvent;
 
 class VirtualKeyboardTest {
 
@@ -18,7 +18,7 @@ class VirtualKeyboardTest {
      */
     @Test
     void testEmptyConstructor(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
+        VirtualKeyboard actual = new VirtualKeyboard();
         assertTrue(actual.getPressedKeys().isEmpty());
         assertTrue(actual.getCharList().isEmpty());
         assertTrue(actual.isParent());
@@ -30,14 +30,14 @@ class VirtualKeyboardTest {
     @Test
     void testSubtickConstructor(){
         Set<Integer> testKeycodeSet = new HashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
 
         List<Character> testCharList = new ArrayList<>();
         testCharList.add('w');
         testCharList.add('s');
 
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, testCharList);
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, testCharList);
 
         assertIterableEquals(testKeycodeSet, actual.getPressedKeys());
         assertIterableEquals(testCharList, actual.getCharList());
@@ -49,10 +49,10 @@ class VirtualKeyboardTest {
      */
     @Test
     void testSetPressedByKeycode(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
-        actual.setPressed(VirtualKey2.W.getKeycode(), true);
+        VirtualKeyboard actual = new VirtualKeyboard();
+        actual.setPressed(VirtualKey.W.getKeycode(), true);
 
-        assertIterableEquals(Arrays.asList(VirtualKey2.W.getKeycode()), actual.getPressedKeys());
+        assertIterableEquals(Arrays.asList(VirtualKey.W.getKeycode()), actual.getPressedKeys());
         assertTrue(actual.isParent());
     }
     
@@ -61,8 +61,8 @@ class VirtualKeyboardTest {
      */
     @Test
     void testFailingSetPressedByKeycode(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
-        actual.setPressed(VirtualKey2.LC.getKeycode(), true);
+        VirtualKeyboard actual = new VirtualKeyboard();
+        actual.setPressed(VirtualKey.LC.getKeycode(), true);
 
         assertTrue(actual.getPressedKeys().isEmpty());
         assertTrue(actual.isParent());
@@ -73,10 +73,10 @@ class VirtualKeyboardTest {
      */
     @Test
     void testSetPressedByKeyname(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
+        VirtualKeyboard actual = new VirtualKeyboard();
         actual.setPressed("W", true);
 
-        assertIterableEquals(Arrays.asList(VirtualKey2.W.getKeycode()), actual.getPressedKeys());
+        assertIterableEquals(Arrays.asList(VirtualKey.W.getKeycode()), actual.getPressedKeys());
         assertTrue(actual.isParent());
     }
 
@@ -86,12 +86,12 @@ class VirtualKeyboardTest {
     @Test
     void testSetUnPressedByKeycode(){
         Set<Integer> testKeycodeSet = new HashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, new ArrayList<>());
-        actual.setPressed(VirtualKey2.W.getKeycode(), false);
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, new ArrayList<>());
+        actual.setPressed(VirtualKey.W.getKeycode(), false);
 
-        assertIterableEquals(Arrays.asList(VirtualKey2.S.getKeycode()), actual.getPressedKeys());
+        assertIterableEquals(Arrays.asList(VirtualKey.S.getKeycode()), actual.getPressedKeys());
     }
 
     /**
@@ -100,12 +100,12 @@ class VirtualKeyboardTest {
     @Test
     void testSetUnPressedByKeyname(){
         Set<Integer> testKeycodeSet = new HashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, new ArrayList<>());
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, new ArrayList<>());
         actual.setPressed("S", false);
 
-        assertIterableEquals(Arrays.asList(VirtualKey2.W.getKeycode()), actual.getPressedKeys());
+        assertIterableEquals(Arrays.asList(VirtualKey.W.getKeycode()), actual.getPressedKeys());
     }
 
     /**
@@ -113,7 +113,7 @@ class VirtualKeyboardTest {
      */
     @Test
     void testAddCharacter(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
+        VirtualKeyboard actual = new VirtualKeyboard();
         actual.addChar('w', false);
 
         assertIterableEquals(Arrays.asList('w'), actual.getCharList());
@@ -125,15 +125,15 @@ class VirtualKeyboardTest {
     @Test
     void testToString(){
         Set<Integer> testKeycodeSet = new LinkedHashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
 
         List<Character> testCharList = new ArrayList<>();
         testCharList.add('w');
         testCharList.add('s');
 
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, testCharList);
-        VirtualKeyboard2 actual2 = new VirtualKeyboard2(testKeycodeSet, new ArrayList<>());
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, testCharList);
+        VirtualKeyboard actual2 = new VirtualKeyboard(testKeycodeSet, new ArrayList<>());
 
         assertEquals("W,S;ws", actual.toString());
         assertEquals("W,S;", actual2.toString());
@@ -144,10 +144,10 @@ class VirtualKeyboardTest {
      */
     @Test
     void testToStringSubticks(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
+        VirtualKeyboard actual = new VirtualKeyboard();
 
-        actual.update(VirtualKey2.W.getKeycode(), true, 'w');
-        actual.update(VirtualKey2.S.getKeycode(), true, 's');
+        actual.update(VirtualKey.W.getKeycode(), true, 'w');
+        actual.update(VirtualKey.S.getKeycode(), true, 's');
 
         assertEquals("W;w\nW,S;s", actual.toString());
     }
@@ -158,15 +158,15 @@ class VirtualKeyboardTest {
     @Test
     void testEquals() {
         Set<Integer> testKeycodeSet = new HashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
 
         List<Character> testCharList = new ArrayList<>();
         testCharList.add('w');
         testCharList.add('s');
 
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, testCharList);
-        VirtualKeyboard2 actual2 = new VirtualKeyboard2(testKeycodeSet, testCharList);
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, testCharList);
+        VirtualKeyboard actual2 = new VirtualKeyboard(testKeycodeSet, testCharList);
         
         assertEquals(actual, actual2);
     }
@@ -177,8 +177,8 @@ class VirtualKeyboardTest {
     @Test
     void testNotEquals() {
         Set<Integer> testKeycodeSet = new HashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
 
         List<Character> testCharList = new ArrayList<>();
         testCharList.add('w');
@@ -191,9 +191,9 @@ class VirtualKeyboardTest {
         List<Character> testCharList3 = new ArrayList<>();
         testCharList3.add('w');
 
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, testCharList);
-        VirtualKeyboard2 test2 = new VirtualKeyboard2(testKeycodeSet, testCharList2);
-        VirtualKeyboard2 test3 = new VirtualKeyboard2(testKeycodeSet, testCharList3);
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, testCharList);
+        VirtualKeyboard test2 = new VirtualKeyboard(testKeycodeSet, testCharList2);
+        VirtualKeyboard test3 = new VirtualKeyboard(testKeycodeSet, testCharList3);
         
         assertNotEquals(actual, test2);
         assertNotEquals(actual, test3);
@@ -206,15 +206,15 @@ class VirtualKeyboardTest {
     @Test
     void testClone() {
         Set<Integer> testKeycodeSet = new HashSet<>();
-        testKeycodeSet.add(VirtualKey2.W.getKeycode());
-        testKeycodeSet.add(VirtualKey2.S.getKeycode());
+        testKeycodeSet.add(VirtualKey.W.getKeycode());
+        testKeycodeSet.add(VirtualKey.S.getKeycode());
 
         List<Character> testCharList = new ArrayList<>();
         testCharList.add('w');
         testCharList.add('s');
 
-        VirtualKeyboard2 actual = new VirtualKeyboard2(testKeycodeSet, testCharList);
-        VirtualKeyboard2 test2 = actual.clone();
+        VirtualKeyboard actual = new VirtualKeyboard(testKeycodeSet, testCharList);
+        VirtualKeyboard test2 = actual.clone();
         
         assertEquals(actual, test2);
     }
@@ -224,16 +224,16 @@ class VirtualKeyboardTest {
      */
     @Test
     void testCopyFrom(){
-    	VirtualKeyboard2 copyFrom = new VirtualKeyboard2();
-    	VirtualKeyboard2 actual = new VirtualKeyboard2();
+    	VirtualKeyboard copyFrom = new VirtualKeyboard();
+    	VirtualKeyboard actual = new VirtualKeyboard();
     	
-    	copyFrom.update(VirtualKey2.W.getKeycode(), true, 'w');
-    	copyFrom.update(VirtualKey2.A.getKeycode(), true, 'a');
+    	copyFrom.update(VirtualKey.W.getKeycode(), true, 'w');
+    	copyFrom.update(VirtualKey.A.getKeycode(), true, 'a');
     	
-    	VirtualKeyboard2 expected = copyFrom.clone();
+    	VirtualKeyboard expected = copyFrom.clone();
     	
-    	actual.update(VirtualKey2.S.getKeycode(), true, 's');
-    	actual.update(VirtualKey2.D.getKeycode(), true, 'd');
+    	actual.update(VirtualKey.S.getKeycode(), true, 's');
+    	actual.update(VirtualKey.D.getKeycode(), true, 'd');
 
         actual.copyFrom(copyFrom);
 
@@ -249,13 +249,13 @@ class VirtualKeyboardTest {
      */
     @Test
     void testUpdate(){
-        VirtualKeyboard2 actual = new VirtualKeyboard2();
-        actual.update(VirtualKey2.W.getKeycode(), true, 'w');
-        actual.update(VirtualKey2.A.getKeycode(), true, 'A');
+        VirtualKeyboard actual = new VirtualKeyboard();
+        actual.update(VirtualKey.W.getKeycode(), true, 'w');
+        actual.update(VirtualKey.A.getKeycode(), true, 'A');
 
-        List<VirtualKeyboard2> expected = new ArrayList<>();
-        expected.add(new VirtualKeyboard2(new HashSet<Integer>(Arrays.asList(VirtualKey2.W.getKeycode())), Arrays.asList('w')));
-        expected.add(new VirtualKeyboard2(new HashSet<Integer>(Arrays.asList(VirtualKey2.W.getKeycode(), VirtualKey2.A.getKeycode())), Arrays.asList('A')));
+        List<VirtualKeyboard> expected = new ArrayList<>();
+        expected.add(new VirtualKeyboard(new HashSet<Integer>(Arrays.asList(VirtualKey.W.getKeycode())), Arrays.asList('w')));
+        expected.add(new VirtualKeyboard(new HashSet<Integer>(Arrays.asList(VirtualKey.W.getKeycode(), VirtualKey.A.getKeycode())), Arrays.asList('A')));
 
         assertIterableEquals(expected, actual.getAll());
     }
@@ -265,11 +265,11 @@ class VirtualKeyboardTest {
      */
     @Test
     void testGetDifference(){
-        VirtualKeyboard2 test = new VirtualKeyboard2(new HashSet<>(Arrays.asList(VirtualKey2.W.getKeycode())), Arrays.asList('w'));
-        VirtualKeyboard2 test2 = new VirtualKeyboard2(new HashSet<>(Arrays.asList(VirtualKey2.W.getKeycode(), VirtualKey2.S.getKeycode())), Arrays.asList('S'));
+        VirtualKeyboard test = new VirtualKeyboard(new HashSet<>(Arrays.asList(VirtualKey.W.getKeycode())), Arrays.asList('w'));
+        VirtualKeyboard test2 = new VirtualKeyboard(new HashSet<>(Arrays.asList(VirtualKey.W.getKeycode(), VirtualKey.S.getKeycode())), Arrays.asList('S'));
         Queue<VirtualKeyboardEvent> actual = new ConcurrentLinkedQueue<>();
         test.getDifference(test2, actual);
-        Queue<VirtualKeyboardEvent> expected = new ConcurrentLinkedQueue<>(Arrays.asList(new VirtualKeyboardEvent(VirtualKey2.S.getKeycode(), true, 'S')));
+        Queue<VirtualKeyboardEvent> expected = new ConcurrentLinkedQueue<>(Arrays.asList(new VirtualKeyboardEvent(VirtualKey.S.getKeycode(), true, 'S')));
 
         assertIterableEquals(expected, actual);
     }
@@ -279,17 +279,17 @@ class VirtualKeyboardTest {
      */
     @Test
     void testGetVirtualEventsPress() {
-    	VirtualKeyboard2 unpressed = new VirtualKeyboard2();
+    	VirtualKeyboard unpressed = new VirtualKeyboard();
     	
-    	VirtualKeyboard2 pressed = new VirtualKeyboard2();
-    	pressed.update(VirtualKey2.W.getKeycode(), true, 'w');
+    	VirtualKeyboard pressed = new VirtualKeyboard();
+    	pressed.update(VirtualKey.W.getKeycode(), true, 'w');
     	
     	// Load actual with the events
     	Queue<VirtualKeyboardEvent> actual = new ConcurrentLinkedQueue<>();
     	unpressed.getVirtualEvents(pressed, actual);
     	
     	// Load expected
-    	List<VirtualKeyboardEvent> expected = Arrays.asList(new VirtualKeyboardEvent(VirtualKey2.W.getKeycode(), true, 'w'));
+    	List<VirtualKeyboardEvent> expected = Arrays.asList(new VirtualKeyboardEvent(VirtualKey.W.getKeycode(), true, 'w'));
     	
     	assertIterableEquals(expected, actual);
     }
@@ -299,17 +299,17 @@ class VirtualKeyboardTest {
      */
     @Test
     void testGetVirtualEventsUnpress() {
-    	VirtualKeyboard2 unpressed = new VirtualKeyboard2();
+    	VirtualKeyboard unpressed = new VirtualKeyboard();
     	
-    	VirtualKeyboard2 pressed = new VirtualKeyboard2();
-    	pressed.update(VirtualKey2.W.getKeycode(), true, 'w');
+    	VirtualKeyboard pressed = new VirtualKeyboard();
+    	pressed.update(VirtualKey.W.getKeycode(), true, 'w');
     	
     	// Load actual with the events
     	Queue<VirtualKeyboardEvent> actual = new ConcurrentLinkedQueue<>();
     	pressed.getVirtualEvents(unpressed, actual);
     	
     	// Load expected
-    	List<VirtualKeyboardEvent> expected = Arrays.asList(new VirtualKeyboardEvent(VirtualKey2.W.getKeycode(), false, Character.MIN_VALUE));
+    	List<VirtualKeyboardEvent> expected = Arrays.asList(new VirtualKeyboardEvent(VirtualKey.W.getKeycode(), false, Character.MIN_VALUE));
     	
     	assertIterableEquals(expected, actual);
     }
@@ -319,9 +319,9 @@ class VirtualKeyboardTest {
      */
     @Test
     void testRepeatEvents(){
-        VirtualKeyboard2 testKb = new VirtualKeyboard2();
+        VirtualKeyboard testKb = new VirtualKeyboard();
 
-        int keycode = VirtualKey2.BACK.getKeycode();
+        int keycode = VirtualKey.BACK.getKeycode();
 
         // Update the keyboard multiple times with the same value
         testKb.update(keycode, true, Character.MIN_VALUE, true);
@@ -330,7 +330,7 @@ class VirtualKeyboardTest {
 
         Queue<VirtualKeyboardEvent> actual = new ConcurrentLinkedQueue<>();
         // Fill "actual" with VirtualKeyboardEvents
-        new VirtualKeyboard2().getVirtualEvents(testKb, actual);
+        new VirtualKeyboard().getVirtualEvents(testKb, actual);
 
         List<VirtualKeyboardEvent> expected = new ArrayList<>();
         // Add expected VirtualKeyboardEvents
@@ -346,9 +346,9 @@ class VirtualKeyboardTest {
      */
     @Test
     void testRepeatEventsFail(){
-        VirtualKeyboard2 testKb = new VirtualKeyboard2();
+        VirtualKeyboard testKb = new VirtualKeyboard();
 
-        int keycode = VirtualKey2.BACK.getKeycode();
+        int keycode = VirtualKey.BACK.getKeycode();
         // Update the keyboard multiple times with the same value.
         testKb.update(keycode, true, Character.MIN_VALUE, false);
         testKb.update(keycode, true, Character.MIN_VALUE, false);
@@ -356,7 +356,7 @@ class VirtualKeyboardTest {
 
         Queue<VirtualKeyboardEvent> actual = new ConcurrentLinkedQueue<>();
         // Fill "actual" with VirtualKeyboardEvents
-        new VirtualKeyboard2().getVirtualEvents(testKb, actual);
+        new VirtualKeyboard().getVirtualEvents(testKb, actual);
 
         List<VirtualKeyboardEvent> expected = new ArrayList<>();
