+++ src/test/java/tasmod/virtual/VirtualMouseTest.java
@@ -15,13 +15,11 @@ import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
+import com.minecrafttas.tasmod.virtual.VirtualMouse;
 import org.junit.jupiter.api.Test;
 
-import com.minecrafttas.tasmod.virtual.VirtualKey2;
-import com.minecrafttas.tasmod.virtual.VirtualKeyboard2;
-import com.minecrafttas.tasmod.virtual.VirtualKeyboardEvent;
-import com.minecrafttas.tasmod.virtual.VirtualMouse2;
-import com.minecrafttas.tasmod.virtual.VirtualMouseEvent;
+import com.minecrafttas.tasmod.virtual.VirtualKey;
+import com.minecrafttas.tasmod.virtual.event.VirtualMouseEvent;
 
 class VirtualMouseTest {
 	
@@ -30,7 +28,7 @@ class VirtualMouseTest {
 	 */
 	@Test
 	void testEmptyConstructor() {
-		VirtualMouse2 actual = new VirtualMouse2();
+		VirtualMouse actual = new VirtualMouse();
 		assertTrue(actual.getPressedKeys().isEmpty());
 		assertEquals(0, actual.getScrollWheel());
 		assertEquals(0, actual.getCursorX());
@@ -44,10 +42,10 @@ class VirtualMouseTest {
 	@Test
 	void testSubtickConstructor() {
         Set<Integer> expected = new HashSet<>();
-        expected.add(VirtualKey2.LC.getKeycode());
-        expected.add(VirtualKey2.RC.getKeycode());
+        expected.add(VirtualKey.LC.getKeycode());
+        expected.add(VirtualKey.RC.getKeycode());
 
-        VirtualMouse2 actual = new VirtualMouse2(expected, -15, 0, 2);
+        VirtualMouse actual = new VirtualMouse(expected, -15, 0, 2);
 
         assertIterableEquals(expected, actual.getPressedKeys());
         assertEquals(-15, actual.getScrollWheel());
@@ -61,10 +59,10 @@ class VirtualMouseTest {
      */
     @Test
     void testSetPressedByKeycode(){
-        VirtualMouse2 actual = new VirtualMouse2();
-        actual.setPressed(VirtualKey2.LC.getKeycode(), true);
+        VirtualMouse actual = new VirtualMouse();
+        actual.setPressed(VirtualKey.LC.getKeycode(), true);
 
-        assertIterableEquals(Arrays.asList(VirtualKey2.LC.getKeycode()), actual.getPressedKeys());
+        assertIterableEquals(Arrays.asList(VirtualKey.LC.getKeycode()), actual.getPressedKeys());
         assertTrue(actual.isParent());
     }
 
@@ -73,10 +71,10 @@ class VirtualMouseTest {
 	 */
 	@Test
 	void testSetPressedByKeyname(){
-		VirtualMouse2 actual = new VirtualMouse2();
+		VirtualMouse actual = new VirtualMouse();
 		actual.setPressed("LC", true);
 
-		assertIterableEquals(Arrays.asList(VirtualKey2.LC.getKeycode()), actual.getPressedKeys());
+		assertIterableEquals(Arrays.asList(VirtualKey.LC.getKeycode()), actual.getPressedKeys());
 		assertTrue(actual.isParent());
 	}
 
@@ -86,12 +84,12 @@ class VirtualMouseTest {
 	@Test
 	void testSetUnPressedByKeycode(){
 		Set<Integer> testKeycodeSet = new HashSet<>();
-		testKeycodeSet.add(VirtualKey2.LC.getKeycode());
-		testKeycodeSet.add(VirtualKey2.MBUTTON9.getKeycode());
-		VirtualMouse2 actual = new VirtualMouse2(testKeycodeSet, 0, 0, 0);
-		actual.setPressed(VirtualKey2.MBUTTON9.getKeycode(), false);
+		testKeycodeSet.add(VirtualKey.LC.getKeycode());
+		testKeycodeSet.add(VirtualKey.MBUTTON9.getKeycode());
+		VirtualMouse actual = new VirtualMouse(testKeycodeSet, 0, 0, 0);
+		actual.setPressed(VirtualKey.MBUTTON9.getKeycode(), false);
 
-		assertIterableEquals(Arrays.asList(VirtualKey2.LC.getKeycode()), actual.getPressedKeys());
+		assertIterableEquals(Arrays.asList(VirtualKey.LC.getKeycode()), actual.getPressedKeys());
 	}
 
 	/**
@@ -100,12 +98,12 @@ class VirtualMouseTest {
 	@Test
 	void testSetUnPressedByKeyname(){
 		Set<Integer> testKeycodeSet = new HashSet<>();
-		testKeycodeSet.add(VirtualKey2.LC.getKeycode());
-		testKeycodeSet.add(VirtualKey2.MBUTTON9.getKeycode());
-		VirtualMouse2 actual = new VirtualMouse2(testKeycodeSet, 0, 0, 0);
+		testKeycodeSet.add(VirtualKey.LC.getKeycode());
+		testKeycodeSet.add(VirtualKey.MBUTTON9.getKeycode());
+		VirtualMouse actual = new VirtualMouse(testKeycodeSet, 0, 0, 0);
 		actual.setPressed("MBUTTON9", false);
 
-		assertIterableEquals(Arrays.asList(VirtualKey2.LC.getKeycode()), actual.getPressedKeys());
+		assertIterableEquals(Arrays.asList(VirtualKey.LC.getKeycode()), actual.getPressedKeys());
 	}
 
 	/**
@@ -114,10 +112,10 @@ class VirtualMouseTest {
 	@Test
 	void testToString(){
 		Set<Integer> testKeycodeSet = new LinkedHashSet<>();
-		testKeycodeSet.add(VirtualKey2.LC.getKeycode());
-		testKeycodeSet.add(VirtualKey2.MC.getKeycode());
+		testKeycodeSet.add(VirtualKey.LC.getKeycode());
+		testKeycodeSet.add(VirtualKey.MC.getKeycode());
 
-		VirtualMouse2 actual = new VirtualMouse2(testKeycodeSet, 10, 100, 120);
+		VirtualMouse actual = new VirtualMouse(testKeycodeSet, 10, 100, 120);
 
 		assertEquals("LC,MC;10,100,120", actual.toString());
 	}
@@ -127,9 +125,9 @@ class VirtualMouseTest {
 	 */
 	@Test
 	void testToStringSubtick(){
-		VirtualMouse2 actual = new VirtualMouse2();
-		actual.update(VirtualKey2.LC.getKeycode(), true, 10, 100, 120);
-		actual.update(VirtualKey2.MC.getKeycode(), true, 0, 12, 3);
+		VirtualMouse actual = new VirtualMouse();
+		actual.update(VirtualKey.LC.getKeycode(), true, 10, 100, 120);
+		actual.update(VirtualKey.MC.getKeycode(), true, 0, 12, 3);
 
 		assertEquals("LC;10,100,120\nLC,MC;0,12,3", actual.toString());
 	}
@@ -140,12 +138,12 @@ class VirtualMouseTest {
 	@Test
 	void testEquals() {
 		Set<Integer> testKeycodeSet = new HashSet<>();
-		testKeycodeSet.add(VirtualKey2.W.getKeycode());
-		testKeycodeSet.add(VirtualKey2.S.getKeycode());
+		testKeycodeSet.add(VirtualKey.W.getKeycode());
+		testKeycodeSet.add(VirtualKey.S.getKeycode());
 
 
-		VirtualMouse2 actual = new VirtualMouse2(testKeycodeSet, -15, 129, 340);
-		VirtualMouse2 actual2 = new VirtualMouse2(testKeycodeSet, -15, 129, 340);
+		VirtualMouse actual = new VirtualMouse(testKeycodeSet, -15, 129, 340);
+		VirtualMouse actual2 = new VirtualMouse(testKeycodeSet, -15, 129, 340);
 
 		assertEquals(actual, actual2);
 	}
@@ -156,20 +154,20 @@ class VirtualMouseTest {
 	@Test
 	void testNotEquals() {
 		Set<Integer> testKeycodeSet = new HashSet<>();
-		testKeycodeSet.add(VirtualKey2.LC.getKeycode());
+		testKeycodeSet.add(VirtualKey.LC.getKeycode());
 
-		VirtualMouse2 actual = new VirtualMouse2(testKeycodeSet, -15, 1, 1);
+		VirtualMouse actual = new VirtualMouse(testKeycodeSet, -15, 1, 1);
 
 		Set<Integer> testKeycodeSet2 = new HashSet<>();
-		testKeycodeSet.add(VirtualKey2.RC.getKeycode());
-		VirtualMouse2 test2 = new VirtualMouse2(testKeycodeSet2, -15, 1, 1);
+		testKeycodeSet.add(VirtualKey.RC.getKeycode());
+		VirtualMouse test2 = new VirtualMouse(testKeycodeSet2, -15, 1, 1);
 
 
 
-		VirtualMouse2 test3 = new VirtualMouse2(testKeycodeSet, -16, 1, 1);
+		VirtualMouse test3 = new VirtualMouse(testKeycodeSet, -16, 1, 1);
 
-		VirtualMouse2 test4 = new VirtualMouse2(testKeycodeSet, -15, 2, 1);
-		VirtualMouse2 test5 = new VirtualMouse2(testKeycodeSet, -15, 1, 2);
+		VirtualMouse test4 = new VirtualMouse(testKeycodeSet, -15, 2, 1);
+		VirtualMouse test5 = new VirtualMouse(testKeycodeSet, -15, 1, 2);
 
 		assertNotEquals(actual, test2);
 		assertNotEquals(actual, test3);
@@ -184,11 +182,11 @@ class VirtualMouseTest {
 	@Test
 	void testClone() {
 		Set<Integer> testKeycodeSet = new HashSet<>();
-		testKeycodeSet.add(VirtualKey2.LC.getKeycode());
-		testKeycodeSet.add(VirtualKey2.MC.getKeycode());
+		testKeycodeSet.add(VirtualKey.LC.getKeycode());
+		testKeycodeSet.add(VirtualKey.MC.getKeycode());
 
-		VirtualMouse2 actual = new VirtualMouse2(testKeycodeSet, 10, 3, 2);
-		VirtualMouse2 test2 = actual.clone();
+		VirtualMouse actual = new VirtualMouse(testKeycodeSet, 10, 3, 2);
+		VirtualMouse test2 = actual.clone();
 
 		assertEquals(actual, test2);
 	}
@@ -198,16 +196,16 @@ class VirtualMouseTest {
 	 */
 	@Test
 	void testCopyFrom() {
-		VirtualMouse2 copyFrom = new VirtualMouse2();
-		VirtualMouse2 actual = new VirtualMouse2();
+		VirtualMouse copyFrom = new VirtualMouse();
+		VirtualMouse actual = new VirtualMouse();
 
-		copyFrom.update(VirtualKey2.LC.getKeycode(), true, 0, 0, 0);
-		copyFrom.update(VirtualKey2.MOUSEMOVED.getKeycode(), false, 120, 10, 20);
+		copyFrom.update(VirtualKey.LC.getKeycode(), true, 0, 0, 0);
+		copyFrom.update(VirtualKey.MOUSEMOVED.getKeycode(), false, 120, 10, 20);
 
-		VirtualMouse2 expected = copyFrom.clone();
+		VirtualMouse expected = copyFrom.clone();
 
-		actual.update(VirtualKey2.MBUTTON12.getKeycode(), true, 0,0,0);
-		actual.update(VirtualKey2.MOUSEMOVED.getKeycode(), true, -120, -10, -10);
+		actual.update(VirtualKey.MBUTTON12.getKeycode(), true, 0,0,0);
+		actual.update(VirtualKey.MOUSEMOVED.getKeycode(), true, -120, -10, -10);
 
 		actual.copyFrom(copyFrom);
 
@@ -227,13 +225,13 @@ class VirtualMouseTest {
 	 */
 	@Test
 	void testUpdate(){
-		VirtualMouse2 actual = new VirtualMouse2();
-		actual.update(VirtualKey2.LC.getKeycode(), true, -30, 118, 42);
-		actual.update(VirtualKey2.MOUSEMOVED.getKeycode(), false, 0, 23, 144);
+		VirtualMouse actual = new VirtualMouse();
+		actual.update(VirtualKey.LC.getKeycode(), true, -30, 118, 42);
+		actual.update(VirtualKey.MOUSEMOVED.getKeycode(), false, 0, 23, 144);
 
-		List<VirtualMouse2> expected = new ArrayList<>();
-		expected.add(new VirtualMouse2(new HashSet<Integer>(Arrays.asList(VirtualKey2.LC.getKeycode())), -30, 118, 42));
-		expected.add(new VirtualMouse2(new HashSet<Integer>(Arrays.asList(VirtualKey2.LC.getKeycode())), 0, 23, 144));
+		List<VirtualMouse> expected = new ArrayList<>();
+		expected.add(new VirtualMouse(new HashSet<Integer>(Arrays.asList(VirtualKey.LC.getKeycode())), -30, 118, 42));
+		expected.add(new VirtualMouse(new HashSet<Integer>(Arrays.asList(VirtualKey.LC.getKeycode())), 0, 23, 144));
 
 		assertIterableEquals(expected, actual.getAll());
 	}
@@ -243,11 +241,11 @@ class VirtualMouseTest {
      */
     @Test
     void testGetDifference(){
-        VirtualMouse2 test = new VirtualMouse2(new HashSet<>(Arrays.asList(VirtualKey2.LC.getKeycode())), 15, 0, 0);
-        VirtualMouse2 test2 = new VirtualMouse2(new HashSet<>(Arrays.asList(VirtualKey2.LC.getKeycode(), VirtualKey2.RC.getKeycode())), 30, 1, 2);
+        VirtualMouse test = new VirtualMouse(new HashSet<>(Arrays.asList(VirtualKey.LC.getKeycode())), 15, 0, 0);
+        VirtualMouse test2 = new VirtualMouse(new HashSet<>(Arrays.asList(VirtualKey.LC.getKeycode(), VirtualKey.RC.getKeycode())), 30, 1, 2);
         Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
         test.getDifference(test2, actual);
-        Queue<VirtualMouseEvent> expected = new ConcurrentLinkedQueue<>(Arrays.asList(new VirtualMouseEvent(VirtualKey2.RC.getKeycode(), true, 30, 1, 2)));
+        Queue<VirtualMouseEvent> expected = new ConcurrentLinkedQueue<>(Arrays.asList(new VirtualMouseEvent(VirtualKey.RC.getKeycode(), true, 30, 1, 2)));
 
         assertIterableEquals(expected, actual);
     }
@@ -257,17 +255,17 @@ class VirtualMouseTest {
      */
     @Test
     void testGetVirtualEventsPress() {
-    	VirtualMouse2 unpressed = new VirtualMouse2();
+    	VirtualMouse unpressed = new VirtualMouse();
     	
-    	VirtualMouse2 pressed = new VirtualMouse2();
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
+    	VirtualMouse pressed = new VirtualMouse();
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
     	
     	// Load actual with the events
     	Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
     	unpressed.getVirtualEvents(pressed, actual);
     	
     	// Load expected
-    	List<VirtualMouseEvent> expected = Arrays.asList(new VirtualMouseEvent(VirtualKey2.LC.getKeycode(), true, 15, 10, 12));
+    	List<VirtualMouseEvent> expected = Arrays.asList(new VirtualMouseEvent(VirtualKey.LC.getKeycode(), true, 15, 10, 12));
     	
     	assertIterableEquals(expected, actual);
     }
@@ -277,17 +275,17 @@ class VirtualMouseTest {
      */
     @Test
     void testGetVirtualEventsUnpress() {
-    	VirtualMouse2 unpressed = new VirtualMouse2();
+    	VirtualMouse unpressed = new VirtualMouse();
     	
-    	VirtualMouse2 pressed = new VirtualMouse2();
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
+    	VirtualMouse pressed = new VirtualMouse();
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
     	
     	// Load actual with the events
     	Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
     	pressed.getVirtualEvents(unpressed, actual);
     	
     	// Load expected
-    	List<VirtualMouseEvent> expected = Arrays.asList(new VirtualMouseEvent(VirtualKey2.LC.getKeycode(), false, 0, 0, 0));
+    	List<VirtualMouseEvent> expected = Arrays.asList(new VirtualMouseEvent(VirtualKey.LC.getKeycode(), false, 0, 0, 0));
     	
     	assertIterableEquals(expected, actual);
     }
@@ -297,11 +295,11 @@ class VirtualMouseTest {
      */
     @Test
     void testSameUpdate() {
-    	VirtualMouse2 unpressed = new VirtualMouse2();
+    	VirtualMouse unpressed = new VirtualMouse();
     	
-    	VirtualMouse2 pressed = new VirtualMouse2();
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
+    	VirtualMouse pressed = new VirtualMouse();
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
     	
     	// Load actual with the events
     	Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
@@ -309,7 +307,7 @@ class VirtualMouseTest {
     	
     	// Load expected
     	List<VirtualMouseEvent> expected = Arrays.asList(
-    			new VirtualMouseEvent(VirtualKey2.LC.getKeycode(), true, 15, 10, 12) // Should only have one keyboard event
+    			new VirtualMouseEvent(VirtualKey.LC.getKeycode(), true, 15, 10, 12) // Should only have one keyboard event
     			); 
     	
     	assertIterableEquals(expected, actual);
@@ -320,11 +318,11 @@ class VirtualMouseTest {
      */
     @Test
     void testScrollWheelDifferent() {
-    	VirtualMouse2 unpressed = new VirtualMouse2();
+    	VirtualMouse unpressed = new VirtualMouse();
     	
-    	VirtualMouse2 pressed = new VirtualMouse2();
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, -30, 10, 12);
+    	VirtualMouse pressed = new VirtualMouse();
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
+    	pressed.update(VirtualKey.LC.getKeycode(), true, -30, 10, 12);
     	
     	// Load actual with the events
     	Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
@@ -332,8 +330,8 @@ class VirtualMouseTest {
     	
     	// Load expected
     	List<VirtualMouseEvent> expected = Arrays.asList(
-    			new VirtualMouseEvent(VirtualKey2.LC.getKeycode(), true, 15, 10, 12),
-    			new VirtualMouseEvent(VirtualKey2.MOUSEMOVED.getKeycode(), false, -30, 10, 12) // Adds an additional "MOUSEMOVED" event with the scroll wheel
+    			new VirtualMouseEvent(VirtualKey.LC.getKeycode(), true, 15, 10, 12),
+    			new VirtualMouseEvent(VirtualKey.MOUSEMOVED.getKeycode(), false, -30, 10, 12) // Adds an additional "MOUSEMOVED" event with the scroll wheel
     			);
     	
     	assertIterableEquals(expected, actual);
@@ -344,11 +342,11 @@ class VirtualMouseTest {
      */
     @Test
     void testCursorXDifferent() {
-    	VirtualMouse2 unpressed = new VirtualMouse2();
+    	VirtualMouse unpressed = new VirtualMouse();
     	
-    	VirtualMouse2 pressed = new VirtualMouse2();
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 11, 12);
+    	VirtualMouse pressed = new VirtualMouse();
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 11, 12);
     	
     	// Load actual with the events
     	Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
@@ -356,8 +354,8 @@ class VirtualMouseTest {
     	
     	// Load expected
     	List<VirtualMouseEvent> expected = Arrays.asList(
-    			new VirtualMouseEvent(VirtualKey2.LC.getKeycode(), true, 15, 10, 12),
-    			new VirtualMouseEvent(VirtualKey2.MOUSEMOVED.getKeycode(), false, 15, 11, 12) // Adds an additional "MOUSEMOVED" event with the cursorX
+    			new VirtualMouseEvent(VirtualKey.LC.getKeycode(), true, 15, 10, 12),
+    			new VirtualMouseEvent(VirtualKey.MOUSEMOVED.getKeycode(), false, 15, 11, 12) // Adds an additional "MOUSEMOVED" event with the cursorX
     			);
     	
     	assertIterableEquals(expected, actual);
@@ -368,11 +366,11 @@ class VirtualMouseTest {
      */
     @Test
     void testCursorYDifferent() {
-    	VirtualMouse2 unpressed = new VirtualMouse2();
+    	VirtualMouse unpressed = new VirtualMouse();
     	
-    	VirtualMouse2 pressed = new VirtualMouse2();
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 12);
-    	pressed.update(VirtualKey2.LC.getKeycode(), true, 15, 10, 120);
+    	VirtualMouse pressed = new VirtualMouse();
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 12);
+    	pressed.update(VirtualKey.LC.getKeycode(), true, 15, 10, 120);
     	
     	// Load actual with the events
     	Queue<VirtualMouseEvent> actual = new ConcurrentLinkedQueue<>();
@@ -380,8 +378,8 @@ class VirtualMouseTest {
     	
     	// Load expected
     	List<VirtualMouseEvent> expected = Arrays.asList(
-    			new VirtualMouseEvent(VirtualKey2.LC.getKeycode(), true, 15, 10, 12),
-    			new VirtualMouseEvent(VirtualKey2.MOUSEMOVED.getKeycode(), false, 15, 10, 120) // Adds an additional "MOUSEMOVED" event with the cursorY
+    			new VirtualMouseEvent(VirtualKey.LC.getKeycode(), true, 15, 10, 12),
+    			new VirtualMouseEvent(VirtualKey.MOUSEMOVED.getKeycode(), false, 15, 10, 120) // Adds an additional "MOUSEMOVED" event with the cursorY
     			);
     	
     	assertIterableEquals(expected, actual);
